EXPDAT-026 added utils showToast, refactored AddFragment, added text check in DetailView, added toast texts to resources [MF,WB]
