+++ app/src/main/java/at/sw21_tug/team_25/expirydates/ui/detailview/ui/DetailView.kt
@@ -2,19 +2,18 @@ package at.sw21_tug.team_25.expirydates.ui.detailview.ui
 
 import android.app.Activity
 import android.app.DatePickerDialog
-import android.content.Context
 import android.content.Context.LAYOUT_INFLATER_SERVICE
 import android.content.Intent
 import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
-import android.view.inputmethod.InputMethodManager
 import android.widget.*
 import at.sw21_tug.team_25.expirydates.MainActivity
 import at.sw21_tug.team_25.expirydates.R
 import at.sw21_tug.team_25.expirydates.data.ExpItem
 import at.sw21_tug.team_25.expirydates.data.ExpItemDao
 import at.sw21_tug.team_25.expirydates.data.ExpItemDatabase
+import at.sw21_tug.team_25.expirydates.utils.Util.Companion.showToast
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import java.time.LocalDate
@@ -54,7 +53,8 @@ class DetailView(private val view: View) : DatePickerDialog.OnDateSetListener {
             val name = popupView.findViewById<TextView>(R.id.product_name)
             val nameEdit = popupView.findViewById<EditText>(R.id.product_name_edit)
             val closePopUpButton = popupView.findViewById<Button>(R.id.closePopUp)
-            val popupBackgroundButton = popupView.findViewById<ImageButton>(R.id.popupBackgroundButton)
+            val popupBackgroundButton =
+                popupView.findViewById<ImageButton>(R.id.popupBackgroundButton)
             val deleteItemButton = popupView.findViewById<Button>(R.id.deleteItem)
             val dateButton = popupView.findViewById<Button>(R.id.exp_date)
             val editButton = popupView.findViewById<Button>(R.id.edit)
@@ -117,9 +117,7 @@ class DetailView(private val view: View) : DatePickerDialog.OnDateSetListener {
             }
 
             editButton.setOnClickListener {
-                is_editable = !is_editable
-                dateButton.isEnabled = is_editable
-                if (is_editable) {
+                if (!is_editable) {
                     editButton.text = activity.getString(R.string.save)
                     closePopUpButton.text = activity.getString(R.string.cancel)
 
@@ -129,12 +127,18 @@ class DetailView(private val view: View) : DatePickerDialog.OnDateSetListener {
                     nameEdit.editableText.clear()
                     nameEdit.editableText.append(name.text)
                 } else {
-                    save(nameEdit.text.toString(), dateButton.text.toString(), activity)
+                    val text = nameEdit.text.toString()
+                    if (text.isEmpty() || text.length > 255) {
+                        showToast(activity, activity.getString(R.string.invalidInput))
+                        return@setOnClickListener
+                    }
+                    save(text, dateButton.text.toString(), activity)
                     cancel(editButton, closePopUpButton, activity, nameEdit, name)
-
-                    name.text = nameEdit.text.toString()
+                    name.text = text
+                    MainActivity.hideKeyboard(activity, popupView)
                 }
-                MainActivity.hideKeyboard(activity, popupView)
+                is_editable = !is_editable
+                dateButton.isEnabled = is_editable
             }
