+++ app/src/main/res/values-ru/strings.xml
@@ -20,4 +20,6 @@
     <string name="exp_asc">срок годности по возрастанию</string>
     <string name="exp_desc">дата истечения срока по убыванию</string>
     <string name="map_label">Ближайшие точки фудшеринга:</string>
+    <string name="invalidInput">Недействительный ввод (должен содержать от 1 до 255 символов)</string>
+    <string name="invalidDate">Недействительная дата (не должна быть прошедшей датой)</string>
 </resources>
