+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,6 @@
     <string name="ru" translatable="false">RU</string>  <!-- Should be international -->
     <string name="language_settings" translatable="false">Language Settings</string>  <!-- Should be international -->
     <string name="map_label">Nearest foodsharing locations:</string>
+    <string name="invalidInput">Invalid Input (Must be between 1 and 255 characters long)</string>
+    <string name="invalidDate">Invalid Date (Must not be a past date)</string>
 </resources>
